Move string conversion functions to @fresha/api-tools-core
